🍗 :: [#349] pose_main_tab_body 리펙토링
Class명 변경, 주석처리 위치 변경, tabNameList는 다른 곳에서 tabName과 한꺼면에 관리하기 위해 삭제
